Fixing build issues for vercel - moved i18n to correct location - removed unused variable warnings
