+++ src/server/trpc/context.ts
@@ -9,7 +9,7 @@ import { prisma } from "../db/client";
 type CreateContextOptions = Record<string, never>;
 
 /** Use this helper for:
- * - testing, so we dont have to mock Next.js' req/res
+ * - testing, so we don't have to mock Next.js' req/res
  * - trpc's `createSSGHelpers` where we don't have req/res
  * @see https://beta.create.t3.gg/en/usage/trpc#-servertrpccontextts
  **/
