+++ src/server/trpc/router/job.ts
@@ -1,9 +1,7 @@
-import { z } from "zod";
-
-import { router, publicProcedure } from "../trpc";
+import { publicProcedure, router } from "../trpc";
 
 export const jobRouter = router({
   all: publicProcedure.query(({ ctx }) => {
     return ctx.prisma.job.findMany();
-  })
+  }),
 });
